Ignore deletion of SMRCTDevServerIP from plist
When rebuilding a non-debug environment, the SMRCTDevServerIP key may not
exist in the plist (because its already been deleted.) This causes PlistBuddy
to exit with a non-0 exit code which fails the build. Its not important if
that fails due to it not existing so this ignore any errors from removing
that key so that the build can continue.